Add high score counter after the first hit

+++ sys/graphics.h
@@ -40,7 +40,7 @@ void DrawRectangle(Rectangle* rect, RGB color, bool fill);
 
 void DrawCollisionSet(CollisionSet* collision_set, RGB color);
 
-void SetTextureAlpha(int texture_id, uint8_t alpha);
+void SetTextureAlpha(int texture_id, double alpha);
 
 RGB HexToRGB(int hex_value);
